==> 01_food_trading/02_unify_country_names.ipynb <==
# Unifying country names

### Trading matrix data:
Country names and codes are in data\FAOSTAT\clean_data\area_country_codes.csv

### country_to_continent data:
I made this data with ChatGPT. Missing "Venezuela" for instance.

### Geopandas countries:
I downloaded the geopandas countries data from https://www.naturalearthdata.com/downloads/10m-cultural-vectors/10m-admin-0-countries-2/


**Will convert everything to lower case and try to match everything to geopandas


import csv
from geopy.distance import great_circle
import pandas as pd
import numpy as np
import geopandas as gpd

fao_countries = pd.read_csv(r"..\data\FAOSTAT\clean_data\fao_country_codes_corrected.csv", encoding="ISO-8859-1")
countries_df = pd.read_csv(r"..\data\FAOSTAT\clean_data\food_trading\country_to_continent.csv") #initial csv done with ChatGPT
geopandas_countries = pd.read_csv(r"..\data\FAOSTAT\clean_data\geopandas_countries.csv")

datasets = [fao_countries, countries_df, geopandas_countries]

for table in datasets:
    print('\n', table.head())

def get_categorical(df):
  """
  Identifies categorical columns in a DataFrame

  Args:
      df (pandas.DataFrame): The DataFrame to identify categorical columns in

  Returns:
      list: A list containing the names of categorical columns
  """
  l_cat = []
  for col in df.columns:
    if df[col].dtype.kind == 'O':  # Check for object dtype (categorical)
      l_cat.append(col)
  return l_cat

# Convert to lower case every object column in all tables
for table in datasets:
    for col in table.columns:
        if col in get_categorical(table):
            table[col] = table[col].astype(str).str.lower()


## I will focus in keeping country codes from fao_countries but only countries in geopandas which is cleaner

fao_countries = fao_countries[fao_countries['Area Code'] < 1000] # This are regions. I exclude from cleaning

fao_list = fao_countries['Area'].tolist()
fao_list.sort()

mycountries_list = countries_df['Country'].tolist()
mycountries_list.sort()

geopandas_list = geopandas_countries['country'].tolist()
geopandas_list.sort()

print(len(fao_list))

print(len(mycountries_list))

print(len(geopandas_list))

#220
#203
#177

for i in fao_list:
    if i not in mycountries_list:
        print(i)

for i in fao_list:
    if i not in geopandas_list:
        print(i)

# Complete myconuntry dataframe to calculate distances

merged_df = pd.merge(countries_df, fao_countries, how='outer', left_on='Country', right_on='Area')


missing = merged_df[merged_df['Country'].isnull() & merged_df['Area'].notnull()]

missing #after completing

**I did not save anything here cause I modified the csv directly**

# Add continents to FAO country codes 

countries_df.head()

fao_countries.head()

merged_df = pd.merge(fao_countries, countries_df, how = 'left', left_on = 'Area', right_on = 'Country')

merged_df.drop(columns = ['Area'], inplace = True)

merged_df.to_csv(r"..\data\FAOSTAT\clean_data\food_trading\country_to_continent_codes.csv", index = False)
