==> 01_food_trading/04_isisland.ipynb <==
# Determine if a country is an island or not

To use it to consider ways of trading

import pandas as pd
import numpy as np

distance_df = pd.read_csv(r"..\data\FAOSTAT\clean_data\food_trading\country_matrix_v1.csv")

distance_df.head()

border_nr_df = distance_df.groupby('country_from')['share_border'].sum().to_frame() #number of border countries (including itself)

border_nr_df.sort_values(by='share_border', ascending=True)

If only shares border with itself = Island or missing info

border_nr_df[border_nr_df['share_border'] == 1]

no_border_countries = border_nr_df[border_nr_df['share_border'] == 1].index.to_list()

island_countries_chatGPT = [ #country islands from chatGPT
    "Antigua and Barbuda",
    "Bahamas",
    "Bahrain",
    "Barbados",
    "Cape Verde (Cabo Verde)",
    "Comoros",
    "Cuba",
    "Cyprus",
    "Dominica",
    "Dominican Republic",
    "East Timor (Timor-Leste)",
    "Fiji",
    "Grenada",
    "Iceland",
    "Indonesia",
    "Ireland",
    "Jamaica",
    "Japan",
    "Kiribati",
    "Madagascar",
    "Maldives",
    "Malta",
    "Marshall Islands",
    "Mauritius",
    "Micronesia (Federated States of Micronesia)",
    "Nauru",
    "New Zealand",
    "Palau",
    "Papua New Guinea",
    "Philippines",
    "Samoa",
    "São Tomé and Príncipe",
    "Seychelles",
    "Singapore",
    "Solomon Islands",
    "Sri Lanka",
    "St. Kitts and Nevis",
    "St. Lucia",
    "St. Vincent and the Grenadines",
    "Tonga",
    "Trinidad and Tobago",
    "Tuvalu",
    "United Kingdom",
    "Vanuatu"
]


island_countries_chatGPT.extend(no_border_countries)
all_islands_ormissing_list = set([item.lower() for item in island_countries_chatGPT])

distance_df['any_island_or_missing'] = np.where(distance_df['country_from'].isin(list(all_islands_ormissing_list)) | 
                                                distance_df['country_to'].isin(list(all_islands_ormissing_list)), 1, 0)

distance_df.to_csv(r"..\data\FAOSTAT\clean_data\food_trading\country_matrix_v2.csv", index=False)
