==> 01_food_trading/05_transport_method.ipynb <==
# Defining transportation method: air, land, water.

## Conditions for aproximating the transportation method:

### Water:

- Importing country is an **island.**
- Importing country continent != partner country continent.

### Land:

- Importing country  **share border** with partner country .
- Same continent.
- Importing country  == partner country --> internal transport

### Air:

- Any country condition but **FOOD IS FRESH** (later)

import pandas as pd
import numpy as np

distance_df = pd.read_csv(r"..\data\FAOSTAT\clean_data\food_trading\country_matrix_v2.csv")

def get_transportation_method(row):
    if (row['any_island_or_missing'] == 1) or (row['same_continent'] == 0):
        return 'water'
    elif row['same_continent'] == 1 or (row['same_border'] == 1) or (row['distance_in_km'] == 0):
        return 'land'

distance_df['transportation_method'] = distance_df.apply(get_transportation_method, axis=1)

distance_df.head()

# Need to correct UK and ireland to be an island.
distance_df[distance_df['country_to'] == 'united kingdom']

distance_df['transportation_method'].value_counts()

# Finally I will add country codes to distance matrix
country_codes = pd.read_csv(r"..\data\FAOSTAT\clean_data\fao_country_codes_corrected.csv", encoding="ISO-8859-1")

country_codes['Area'] = country_codes['Area'].str.lower()

distance_df = distance_df.merge(country_codes, how="left", left_on="country_from", right_on="Area")
distance_df = distance_df.rename(columns={'Area Code': 'country_from_code'})
distance_df.drop(columns='Area', inplace=True)

distance_df = distance_df.merge(country_codes, how="left", left_on="country_to", right_on="Area")
distance_df = distance_df.rename(columns={'Area Code': 'country_to_code'})
distance_df.drop(columns='Area', inplace=True)

distance_df['country_from_code'] = distance_df['country_from_code'].fillna(-1).astype(int)
distance_df['country_to_code'] = distance_df['country_to_code'].fillna(-1).astype(int)

distance_df.head()

distance_df.to_csv(r"..\data\FAOSTAT\clean_data\food_trading\country_matrix_v3.csv", index=False)
