==> 01_food_trading/07_merge_trade_mx_with_distance.ipynb <==
import pandas as pd
import numpy as np

#save and load as parquet (for big files)
import pyarrow as pa
import pyarrow.parquet as pq

# I drop rows with 0 value of trading for this v2
trade_mx = pd.read_parquet(r"..\data\FAOSTAT\clean_data\food_trading\trade_matrix__imports_clean_codesOnly_v2.parquet")

distance_df = pd.read_csv(r"..\data\FAOSTAT\clean_data\food_trading\country_matrix_v4_final.csv")

distance_df.head()

trade_mx.head()

# Convert variables type

To decrease memory usage

#Convert continent to category
cols = ['continent_from', 'Continent_to']
for col in cols:
    distance_df[col] = distance_df[col].astype('category')

# Clean trade matrix
## to make it smaller

Create keys to merge

distance_df['key'] = distance_df['country_from_code'].astype(str) + '-' + distance_df['country_to_code'].astype(str)
trade_mx['key'] = trade_mx['Reporter Country Code'].astype(str) + '-' + trade_mx['Partner Country Code'].astype(str)

## I have to drop countries that don`t exist in trade matrix cause results where not ok. I had mixed continents.

I drop NaN in "share_border" and "-1" in contry codes from distance_df

distance_df.dropna(subset=['share_border'], inplace=True)

distance_df = distance_df[(distance_df['country_from_code'] > 0) & (distance_df['country_to_code'] > 0)]

full_trade_mx = trade_mx.merge(distance_df, on='key', how='inner')

full_trade_mx.head()

cols = ['country_from_code', 'country_to_code', 'internal_distance', 'key', 'country_from', 'country_to', 'continent_from', 'Continent_to']
full_trade_mx.drop(columns=cols, inplace=True)

full_trade_mx.to_parquet(r"..\data\FAOSTAT\clean_data\food_trading\trade_matrix__imports_clean_distance_v3.parquet", index=False)
